Python/Italiano/Data_Horizon/etl_data_erp/requirements.txt
Flask~=3.1.0
Office365-REST-Python-Client==2.5.13
pg8000==1.31.2
SQLAlchemy==2.0.36
python-dotenv==1.0.1
pandas==2.2.3